updates to eval scripts

+++ convert_data.py
@@ -26,12 +26,14 @@ def dnerf_to_dynaGS(inp_folder, out_folder, folder='train', reference_fname=None
         # 'w2c': 150 x 27 x (4x4 matrix)
         # 'fn': 150 x 27 filenames: '1/000000.jpg'.,...
         # data['cam_id']: 150 x 27
+    
+    os.makedirs(join(out_folder, 'train'), exist_ok=True)
 
     if folder == 'train':
-        assert os.path.exists(join(inp_folder, folder, 'init_pt_cld.npz')), "no init_pt_cld.npz"
-        os.makedirs(join(out_folder, 'train'), exist_ok=True)
-        shutil.copy(
-            join(inp_folder, folder, 'init_pt_cld.npz'), join(out_folder, 'train', 'init_pt_cld.npz'))
+        #assert os.path.exists(join(inp_folder, folder, 'init_pt_cld.npz')), "no init_pt_cld.npz"
+        if os.path.exists(join(inp_folder, folder, 'init_pt_cld.npz')):
+            shutil.copy(
+                join(inp_folder, folder, 'init_pt_cld.npz'), join(out_folder, 'train', 'init_pt_cld.npz'))
     new_meta_data = dict()
     imgs = natsorted(glob(join(inp_folder, folder, "*.png"))) # r_camid_tstep.png 
     tsteps, cam_ids = [], []
@@ -241,25 +243,29 @@ def dynaGS_to_dnerf(
 
     print(f"Done writing to {new_meta_fname}")
     return tot_time
-### OPTION1: convert basketball data to dnerf format    
-# inp_folder = "data/corl_1_dense"
-# out_folder = "data/corl_1_dense_pano" 
-# os.makedirs(out_folder, exist_ok=True)
-# cam_id_offset = 0
-# for folder in ["train", "test"]:
-#     cam_id_offset = dnerf_to_dynaGS(inp_folder, out_folder, folder, cam_id_offset=cam_id_offset)
-#     print(f"cam_id_offset: {cam_id_offset}")
+### OPTION1: convert basketball data to dnerf format  
+
+input_folders = glob("data/final_scenes_bg/*")
+output_folders = [f.replace("final_scenes_bg", "final_scenes_bg_dnerf") for f in input_folders]
+
+for inp_folder, out_folder in zip(input_folders, output_folders):
+    os.makedirs(out_folder, exist_ok=True)
+    cam_id_offset = 0
+    print(f"converting {inp_folder} to {out_folder}")
+    for folder in ["train", "test"]:
+        cam_id_offset = dnerf_to_dynaGS(inp_folder, out_folder, folder, cam_id_offset=cam_id_offset)
+        print(f"cam_id_offset: {cam_id_offset}")
 
 #### OPTION2: convert dnerf data to dynamic3dgaussians format
-inp_folder = "data/basketball"
-out_folder = "data/basketball_dnerf"
+#inp_folder = "data/basketball"
+#out_folder = "data/basketball_dnerf"
 
-#if out folder exists, destroy it
-if os.path.exists(out_folder):
-    shutil.rmtree(out_folder)
+##if out folder exists, destroy it
+#if os.path.exists(out_folder):
+    #shutil.rmtree(out_folder)
 
-# for folder in ["train", "test"]:
-train_tot_time = dynaGS_to_dnerf(inp_folder, out_folder, "train", mask_seg=False)
-_ = dynaGS_to_dnerf(inp_folder, out_folder, "test", mask_seg=False,tot_time=train_tot_time) # to make sure a subsampled test set doesn't cause problems
+## for folder in ["train", "test"]:
+#train_tot_time = dynaGS_to_dnerf(inp_folder, out_folder, "train", mask_seg=False)
+#_ = dynaGS_to_dnerf(inp_folder, out_folder, "test", mask_seg=False,tot_time=train_tot_time) # to make sure a subsampled test set doesn't cause problems
 
 # breakpoint()
